外部受入内で整数をIntegerで変換しようとしてるとこがあったため修正

+++ nts.uk/uk.com/exio.ctx/nts.uk.ctx.exio.dom/src/main/java/nts/uk/ctx/exio/dom/input/canonicalize/domains/generic/IndependentCanonicalization.java
@@ -200,7 +200,7 @@ public abstract class IndependentCanonicalization implements DomainCanonicalizat
 				value = stringified.asString();
 				break;
 			case INT:
-				value = stringified.asInteger();
+				value = stringified.asLong();
 				break;
 			case REAL:
 				value = stringified.asBigDecimal();
